fix(dts): order remapped array mapped surfaces
