+++ crates/tsz-solver/src/type_queries/mod.rs
@@ -36,6 +36,7 @@ pub mod flow;
 pub mod iterable;
 pub mod mapped;
 pub mod mapped_declaration_surface;
+pub mod mapped_display_order;
 pub mod shape_queries;
 pub mod traversal;
 
@@ -93,6 +94,8 @@ pub use flow::*;
 pub use iterable::*;
 pub use mapped::*;
 pub use mapped_declaration_surface::*;
+pub use mapped_display_order::*;
+pub use mapped_display_order::*;
 pub use shape_queries::*;
 pub use traversal::*;
